+++ .gitignore
@@ -20,3 +20,8 @@
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here http://doc.crates.io/guide.html#cargotoml-vs-cargolock
 Cargo.lock
+
+# Certificate files
+*.der
+*.pem
+*.key
