+++ src/lib.rs
@@ -3,15 +3,11 @@ extern crate solicit;
 extern crate rustc_serialize;
 extern crate time;
 extern crate openssl;
+extern crate btls;
 
-pub mod provider;
+pub mod client;
 pub mod notification;
 pub mod payload;
 pub mod device_token;
+pub mod apns_token;
 pub mod response;
-
-pub use provider::{Provider, AsyncResponse};
-pub use notification::{Notification, NotificationOptions};
-pub use payload::{APS, APSAlert, APSLocalizedAlert, Payload, CustomData};
-pub use device_token::DeviceToken;
-pub use response::{Response, APNSError, APNSStatus};
